* Spec v1.3.1.0 ==> have completed

+++ src/main/java/com/mixram/telegram/bot/services/domain/entity/Chat.java
@@ -28,7 +28,7 @@ public class Chat implements TelegramApiEntity {
      * @since 0.1.0.0
      */
     @JsonProperty("id")
-    private Long chatId;
+    private Integer chatId;
     /**
      * Type of chat, can be either “private”, “group”, “supergroup” or “channel”.
      *
